RAG_local/parent_document/requirements.txt
openai>=0.27.0
pinecone-client>=2.2.1
tiktoken>=0.3.3
numpy>=1.21.6